refactoring all explorer views to do server side rendering, other interface tweaks
