+++ yeastphenome/apps/common/templates/base/navigation.html
@@ -74,7 +74,7 @@ ul.nav li.dropdown:hover ul.dropdown-menu {
     <!-- Navigation -->
     <div class="collapse navbar-collapse global-navbar" id="global-navbar">
       <ul class="nav navbar-nav navbar-right">
-        <li class="dropdown">
+        <!--<li class="dropdown">
           <a href="{% url 'common:getting-started' %}" {% if active == "getting-started" %}style="background-color:rgb(177, 103, 24)"{% endif %} role="button" aria-haspopup="true" aria-expanded="false">Getting Started <i class="fas fa-caret-down" aria-hidden="true"></i></a>
           <ul class="dropdown-menu dropdown-menu-left">
             <li><a href="{% url 'common:introduction' %}">Introduction</a></li>
@@ -82,7 +82,7 @@ ul.nav li.dropdown:hover ul.dropdown-menu {
             <li><a href="{% url 'common:tutorials' %}">Tutorials</a></li>
             <li><a href="{% url 'common:advanced' %}">Advanced</a></li>
           </ul>
-        </li>
+        </li>-->
         <li class="dropdown">
           <a href="{% url 'common:explorer' %}" {% if active == "explorer" %}style="background-color:rgb(177, 103, 24)"{% endif %} role="button" aria-haspopup="true" aria-expanded="false">Explore Data <i class="fas fa-caret-down" aria-hidden="true"></i></a>
           <ul class="dropdown-menu dropdown-menu-left">
@@ -98,12 +98,11 @@ ul.nav li.dropdown:hover ul.dropdown-menu {
         <li class="dropdown">
           <a {% if active == "about"%}style="background-color:rgb(177, 103, 24)"{% endif %} role="button" aria-haspopup="true" data-toggle="dropdown" aria-expanded="false">About <i class="fas fa-caret-down" aria-hidden="true"></i></a>
           <ul class="dropdown-menu dropdown-menu-left">
-            <li><a href="{% url 'common:about' %}">About</a></li>
-            <li style="margin: 8px 0" role="separator" class="divider"></li>
-            <li><a href="{% url 'papers:contributors' %}">Paper Contributors</a></li>
-            <li><a href="{% url 'common:contributors' %}">Contributors</a></li>
-            <li><a href="{% url 'common:faq' %}">Frequently Asked Questions</a></li>
+            <li><a href="{% url 'common:about' %}">Project</a></li>
+            <li><a href="{% url 'common:contributors' %}">Authors</a></li>
+            <li><a href="{% url 'common:data_contributors' %}">Data Contributors</a></li>
             <li><a href="{% url 'common:stats' %}">Stats</a></li>
+            <li><a href="{% url 'common:faq' %}">Frequently Asked Questions</a></li>
           </ul>
         </li>
        {% if GITHUB_REPOSITORY %}<li><a href="{{ GITHUB_REPOSITORY }}/issues" target="_blank" rel="noopener"><i class="fas fa-comments visible-xs-inline"></i> Support</a></li>{% endif %}
