+++ yeastphenome/apps/common/templates/graphs/dataset-genes/quantitative.html
@@ -65,9 +65,11 @@ $.each(data, function(i, e){
 // Create a lookup for gene values
 var gene_values_lookup = Object()
 var gene_rank_lookup = Object()
+var gene_id_lookup = Object()
 $.each(gene_values, function(i, e){
     gene_values_lookup[e.label] = e.value
     gene_rank_lookup[e.label] = e.rank
+    gene_id_lookup[e.label] = e.gene_id
 })
 
 var yMax = d3.max(data, function(d){return d.length});
@@ -143,6 +145,7 @@ update_plot = function(d, i) {
         // Grab the value and pvalue
         gene_value = gene_values_lookup[e] || ""
         gene_rank = gene_rank_lookup[e] || ""
+        gene_id = gene_id_lookup[e] || ""
 
         // If we have aliases and common names, add them too
         has_aliases = false
@@ -150,14 +153,14 @@ update_plot = function(d, i) {
         if (e in common_names) {
             $.each(common_names[e], function(ii, ee){
                 gene_value = gene_values_lookup[e]
-                content += "<tr><td id='td-" + ee + "'><a href='/genes?query="+ e + "'>" + e + " / " + ee + "</a></td><td>" + gene_value + "</td><td>" + gene_rank + "</td></tr>\n"
+                content += "<tr><td id='td-" + ee + "'><a href='/genes/"+ gene_id + "/'>" + e + " / " + ee + "</a></td><td>" + gene_value + "</td><td>" + gene_rank + "</td></tr>\n"
             })
            has_aliases = true
         }
 
         // Only add the gene if it doesn't have aliases
         if (!has_aliases){
-            content += "<tr><td id='td-" + e + "'><a href='/genes?query="+ e + "'>" + e + " / " + e + "</a></td><td>" + gene_value + "</td><td>" + gene_rank + "</td></tr>\n"
+            content += "<tr><td id='td-" + e + "'><a href='/genes/"+ gene_id + "/'>" + e + " / " + e + "</a></td><td>" + gene_value + "</td><td>" + gene_rank + "</td></tr>\n"
         }
     })
