+++ yeastphenome/apps/common/templates/main/contributors.html
@@ -37,7 +37,7 @@
     </div>
 </div>
 <div class="row" style="padding-top:20px">
-        <p class="alert alert-warning">A very special thanks goes to our <a href="{% url 'papers:contributors' %}">papers contributors</a>!</p>
+        <p class="alert alert-warning">A very special thanks goes to our <a href="{% url 'common:data_contributors' %}">data contributors</a>!</p>
 </div>
 </div>
 <!-- Right sidebar, goes away for small screens -->
@@ -61,7 +61,7 @@
             Programmers
         </a>
         <a href="#papers" class="list-group-item">
-            Papers Contributors
+            Data Contributors
         </a>
     </div>
 </div>
