+++ yeastphenome/apps/conditions/templates/conditions/tag_browser.html
@@ -15,7 +15,7 @@
 		</thead>
 		<tbody id="papersTableBody">
 		{% for tag in tags %}<tr>
-		    <td><a href="/conditions/tags/{{ tag.id }}">{{ tag.name }}</a></td>
+		    <td><a href="/conditions/?tags={{ tag.name }}">{{ tag.name }}</a></td>
 		    <td>{{ tag.conditiontype_set.count }}</td>
           	</tr>{% endfor %}
 		</tbody>
