+++ yeastphenome/apps/papers/urls.py
@@ -7,7 +7,6 @@ urlpatterns = [
     path("", views.paper_explorer, name="all"),
     path("year/<int:year>/", views.paper_explorer, name="all_year"),
     path("<int:pk>/", views.PaperDetailView.as_view(), name="detail"),
-    path("contributors/", views.ContributorsListView.as_view(), name="contributors"),
     # Data
     path(
         "<int:paper_id>/%s_<int:pmid>_datasets_list.txt" % settings.DOWNLOAD_PREFIX,
