+++ yeastphenome/apps/phenotypes/templates/phenotypes/phenotypes_table.html
@@ -1,7 +1,7 @@
 {% load my_filters %}
 <div class="row">
     <div class="col-md-12">
-        {% if observables %}<table id="papersTable" class="tablesorter table table-condensed" width="100%">
+        <table id="papersTable" class="tablesorter table table-condensed" width="100%">
             <thead>
             <tr>
                 <th>Name</th>
@@ -18,6 +18,5 @@
         </tr>{% endfor %}
         </tbody>
         </table>
-        {% else %}<p class="alert alert-info">There are no phenotypes in the database.</p>{% endif %}
     </div>
 </div>
